pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "non-tech-commit-tracker"
version = "0.0.1"
description = "non-tech commit tracker"
readme = "README.md"
requires-python = ">=3.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "notion-client",
    "python-dotenv"
]
